% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_user_about_data_cleaning.R
\name{ask_user_about_data_cleaning}
\alias{ask_user_about_data_cleaning}
\title{ask_user_about_data_cleaning}
\usage{
ask_user_about_data_cleaning()
}
\value{
a boolean value TRUE for yes / FALSE for no
}
\description{
prompt a question to ask the user about performing data cleaning or not
}
\details{
display a yes/no question and take a string answer as input.
}
