% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_user_about_data_cleaning_5_perCent.R
\name{ask_user_about_data_cleaning_5_perCent}
\alias{ask_user_about_data_cleaning_5_perCent}
\title{ask_user_about_data_cleaning_5_perCent}
\usage{
ask_user_about_data_cleaning_5_perCent()
}
\value{
a boolean value TRUE for yes / FALSE for no
}
\description{
prompt a question to ask the user about performing data cleaning_5\% or not
}
\details{
display a yes/no question and take a string answer as input
}
