% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_user_about_which_dataset_to_use.R
\name{ask_user_about_which_dataset_to_use}
\alias{ask_user_about_which_dataset_to_use}
\title{Ask the user about which dataset to use}
\usage{
ask_user_about_which_dataset_to_use()
}
\value{
a boolean value TRUE for yes / FALSE for no
}
\description{
prompt a question to ask weather to select the user's own dataset or directly use the default Drosophila melanogaster R6 pre-loaded dataset.
}
\details{
display a yes/no question and take a string answer as input
}
