% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brec_chromosome.R
\name{Brec_chromosome}
\alias{Brec_chromosome}
\title{Brec_chromosome}
\usage{
Brec_chromosome(genomeName = "genome name", inputData, inputChrID,
  spanFromGui)
}
\arguments{
\item{genomeName}{string of the name of the genome (organism) forwhich the data is provided. This will be used for output and plot file names. Default value is an empty string "".}

\item{inputData}{......}

\item{inputChrID}{numeric or character value which identifies the chromosome to be processed no default value is available.}
}
\value{
BREC_chromosome_results_list
}
\description{
The main function that initiates the workflow of recombination rates and chromatin boundaries estimation along a chromosme
}
\details{
This function applies Brec on a chromosome by taking as parameters the name of the input genome and the chromosome number (or ID)
in addition of some other optional parameters like
}
