% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brec_chromosome_part_2.R
\name{Brec_chromosome_part_2}
\alias{Brec_chromosome_part_2}
\title{Brec_chromosome_part_2}
\usage{
Brec_chromosome_part_2(BREC_chromosome_results_list, chrType_object)
}
\arguments{
\item{BREC_chromosome_results_list}{...}

\item{chrType_object}{..}
}
\value{
BrecResultsList
}
\description{
The main function that initiates the workflow of recombination rates and chromatin boundaries estimation along a chromosme
}
\details{
This function applies Brec on a chromosome by taking as parameters the name of the input genome and the chromosome number (or ID)
in addition of some other optional parameters like
}
