% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_5_perCent_chromosome_data.R
\name{clean_5_perCent_chromosome_data}
\alias{clean_5_perCent_chromosome_data}
\title{clean_5_perCent_chromosome_data}
\usage{
clean_5_perCent_chromosome_data(chromosome, genomeName, chrID)
}
\arguments{
\item{chromosome}{xxx}

\item{genomeName}{xxx}

\item{chrID}{xxx}
}
\value{
cleanedChromosome_5_perCent
}
\description{
clean chromosome data by deleting 5\% of the detected outliers
}
