% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chromosome_data.R
\name{clean_chromosome_data}
\alias{clean_chromosome_data}
\title{clean_chromosome_data}
\usage{
clean_chromosome_data(chromosome, genomeName, chrID)
}
\arguments{
\item{chromosome}{xxx}

\item{genomeName}{xxx}

\item{chrID}{xxx}
}
\value{
cleanedChromosome
}
\description{
clean chromosome data by deleting all detected outliers
}
