% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cumulated_R_squared_2directions.R
\name{compute_cumulated_R_squared_2directions}
\alias{compute_cumulated_R_squared_2directions}
\title{compute_cumulated_R_squared_2directions}
\usage{
compute_cumulated_R_squared_2directions(chromosome)
}
\arguments{
\item{chromosome}{xx}
}
\value{
R2DataFrame2D
}
\description{
compute cumulated R-squared vectors in two directions : starting from the beginning then the end of chromosome
}
