% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sliding_window_size.R
\name{compute_sliding_window_size}
\alias{compute_sliding_window_size}
\title{compute_sliding_window_size}
\usage{
compute_sliding_window_size(chromosome)
}
\arguments{
\item{chromosome}{xx}
}
\value{
maxEcarts is the double value representing the size of the sliding window
}
\description{
compute sliding window size to be used for chrmatin boundaries extraction
}
