% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_density_test.R
\name{data_density_test}
\alias{data_density_test}
\title{data_density_test}
\usage{
data_density_test(testChromosome)
}
\arguments{
\item{testChromosome}{the chromosome data to be tested}
}
\value{
boolean value saying weather data density is good/TRUE or not/FALSE
}
\description{
test data density
}
