% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_distribution_test.R
\name{data_distribution_test}
\alias{data_distribution_test}
\title{data_distribution_test}
\usage{
data_distribution_test(testChromosome)
}
\arguments{
\item{testChromosome}{the chromosome data to be tested}
}
\value{
boolean value saying weather data distribution is good (TRUE) or not (FALSE)
}
\description{
test data distribution using a chi-squared test with a multivariate normal distribution
}
