% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_test.R
\name{data_quality_test}
\alias{data_quality_test}
\title{data_quality_test}
\usage{
data_quality_test(testChromosome)
}
\arguments{
\item{testChromosome}{xx}
}
\value{
goodDataQuality a boolean value
}
\description{
test data quality based on markers number/density and distribution
}
