% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{DataAeAegypti3cols_OriginalFile}
\alias{DataAeAegypti3cols_OriginalFile}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{http://science.sciencemag.org/content/356/6333/92}
}
\author{
Olga Dudchenko, Erez Lieberman Aiden \email{erez@math.harvard.edu}
}
\keyword{Aedes_aegypti}
\keyword{data}
\keyword{mosquito}
