% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{dmel}
\alias{dmel}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{https://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.1002112}
}
\author{
Russell B. Corbett-Detig, Timothy B. Sackton \email{tsackton@oeb.harvard.edu}
}
\keyword{Drosophila_melanogaster}
\keyword{data}
\keyword{fruitfly}
