% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{Dmel_R6_formatted_v2}
\alias{Dmel_R6_formatted_v2}
\title{This is data to be included in my package}
\description{
A dataset containing the markers (genes) ... The attributes are as follows:
}
\details{
\itemize{
  \item gene diamond colour, from J (worst) to D (best)
  \item symbol.  how clear the diamond is (I1 (worst), SI1, SI2, VS1, VS2, VVS1, VVS2, IF (best))
  \item chr. chromosome ID (X, 2, 3)
  \item cm a measurement of genetic distance of the gene on this chromosome  (>=0)
  \item mb. a measurement of physical distance of the gene on this chromosome (>=0)
}
}
\references{
\url{https://annasfistonlavier.com/}
}
\author{
Anna-Sophie Fiston-Lavier \email{anna-sophie.fiston-lavier@umontpellier.fr}
}
\keyword{Drosophila_melanogaster_R6}
\keyword{data}
\keyword{fruitfly}
