% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_recombination_rates.R
\name{estimate_recombination_rates}
\alias{estimate_recombination_rates}
\title{estimate_recombination_rates}
\usage{
estimate_recombination_rates(chromosome)
}
\arguments{
\item{chromosome}{xx}
}
\value{
RR_object
}
\description{
estimate recombination rates according to the MareyMap method using a 3rd degree polynomial that correlates between genetic and physical maps of a chromosome
}
