% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_recombination_rates_loess.R
\name{estimate_recombination_rates_loess}
\alias{estimate_recombination_rates_loess}
\title{estimate_recombination_rates_loess}
\usage{
estimate_recombination_rates_loess(chromosome, spanVal)
}
\arguments{
\item{chromosome}{xx}

\item{spanVal}{xx}
}
\value{
RR_object
}
\description{
estimate recombination rates according to the MareyMap method using a 3rd degree polynomial that correlates between genetic and physical maps of a chromosome
}
