% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CB.R
\name{extract_CB}
\alias{extract_CB}
\title{extract chromatin boundaries}
\usage{
extract_CB(chromosome, RR_object, R2DataFrame2D, chrID, chrType,
  minRR_object)
}
\arguments{
\item{chromosome}{xx}

\item{RR_object}{xx}

\item{R2DataFrame2D}{xxx}

\item{chrID}{xxx}

\item{chrType}{xxx}

\item{minRR_object}{xxx}
}
\value{
heteroChromatinBoundaries a dataframe of both centromeric chromatinboundaries, values are physical locations
}
\description{
approch : start from min(RR) and go in both ways on R2 plots with a sliding window
            for which the size is computed automatically based on physical_map
            range (écart) of each 2 consecutive points. So, the value of the maximum
            range will represent the sliding window size
}
