% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CB_with_centromeric_gap.R
\name{extract_CB_with_centromeric_gap}
\alias{extract_CB_with_centromeric_gap}
\title{extract_CB_with_centromeric_gap}
\usage{
extract_CB_with_centromeric_gap(chromosome, RR_object, R2DataFrame2D,
  chrID, chrType, minRR_object)
}
\arguments{
\item{chromosome}{xx}

\item{RR_object}{xx}

\item{R2DataFrame2D}{xx}

\item{chrID}{xx}

\item{chrType}{....}

\item{minRR_object}{xxx}
}
\value{
heteroChromatinBoundaries a dataframe of both centromeric chromatinboundaries, values are physical locations
}
\description{
extract chromatin boundaries when a chromosome has a centromeric gap
}
\details{
this function concerns cases with a centromeric gap
}
