% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CB_with_no_centromeric_gap_using_RR.R
\name{extract_CB_with_no_centromeric_gap_using_RR}
\alias{extract_CB_with_no_centromeric_gap_using_RR}
\title{extract_CB_with_no_centromeric_gap_using_RR}
\usage{
extract_CB_with_no_centromeric_gap_using_RR(chromosome, RR_object,
  R2DataFrame2D, chrID, chrType)
}
\arguments{
\item{chromosome}{xx}

\item{RR_object}{xx}

\item{R2DataFrame2D}{xx}

\item{chrID}{xx}

\item{chrType}{....}
}
\value{
heteroChromatinBoundaries a dataframe of both centromeric chromatinboundaries, values are physical locations
}
\description{
extract chromatin boundaries when a chromosome has no centromeric gap
}
