% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_telomeres_boundaries.R
\name{extract_telomeres_boundaries}
\alias{extract_telomeres_boundaries}
\title{extract_telomeres_boundaries}
\usage{
extract_telomeres_boundaries(chromosome, R2DataFrame2D, chrID, chrType,
  minRR_object)
}
\arguments{
\item{chromosome}{xxx}

\item{R2DataFrame2D}{xxx}

\item{chrID}{xxx}

\item{chrType}{....}

\item{minRR_object}{xxx}
}
\value{
telomeres_boundaries
}
\description{
identify telomeres boundaries to use later for extrapolating the RR plot, using min R2 in both directions
}
\details{
goal
}
