% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_RR_estimates.R
\name{extrapolate_RR_estimates}
\alias{extrapolate_RR_estimates}
\title{extrapolate_RR_estimates}
\usage{
extrapolate_RR_estimates(chromosome, RR_object, heteroChromatinBoundaries,
  telomeres_boundaries, chrID, chrType, minRR_object)
}
\arguments{
\item{chromosome}{xxx}

\item{RR_object}{xxx}

\item{heteroChromatinBoundaries}{xxx}

\item{telomeres_boundaries}{xxx}

\item{chrID}{xxx}

\item{chrType}{....}

\item{minRR_object}{...}
}
\value{
RR_object
}
\description{
extrapolate recombination rate values in heterochromatin regions along the chromosome
}
\details{
two cases to handle : centromere and telomeres
goal : set recombination rate values to zero in both cases (since theses genomic regions are known for a null recombination)
}
