% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_HCB_output_data_file.R
\name{generate_HCB_output_data_file}
\alias{generate_HCB_output_data_file}
\title{generate_HCB_output_data_file}
\usage{
generate_HCB_output_data_file(genomeName, chrID, chromosome,
  heteroChromatinBoundaries, telomeres_boundaries, output_path, chrType)
}
\arguments{
\item{genomeName}{xxx}

\item{chrID}{xxx}

\item{chromosome}{xx}

\item{heteroChromatinBoundaries}{xx}

\item{telomeres_boundaries}{xx}

\item{output_path}{xx}

\item{chrType}{....}
}
\value{
RR_fileContent
}
\description{
this function generates and saves the output as a .txt data file for heterochromatin boundaries with centromere and telomeres estimated sizes
}
