% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RR_output_data_file.R
\name{generate_RR_output_data_file}
\alias{generate_RR_output_data_file}
\title{generate_RR_output_data_file}
\usage{
generate_RR_output_data_file(genomeName, chrID, chromosome, RR_object,
  output_path)
}
\arguments{
\item{genomeName}{xxx}

\item{chrID}{xxx}

\item{chromosome}{xx}

\item{RR_object}{xx}

\item{output_path}{xx}
}
\value{
RR_fileContent
}
\description{
this function generates and saves the output as a .txt data file for recombination rate estimates along a chromosome
}
