% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_from_Arms.R
\name{get_chromosome_from_Arms}
\alias{get_chromosome_from_Arms}
\title{get_chromosome_from_Arms}
\usage{
get_chromosome_from_Arms(inputData, chrID)
}
\arguments{
\item{inputData}{xx}

\item{chrID}{xxx}
}
\value{
chromosome
}
\description{
...
}
