% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_from_inputData.R
\name{get_chromosome_from_inputData}
\alias{get_chromosome_from_inputData}
\title{get_chromosome_from_inputData}
\usage{
get_chromosome_from_inputData(inputData, chrID)
}
\arguments{
\item{inputData}{xx}

\item{chrID}{xx}
}
\value{
chromosome : a dataframe of two columns cm and mp
}
\description{
get the data of specified chromosome from the input dataset
}
