% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_type.R
\name{get_chromosome_type}
\alias{get_chromosome_type}
\title{get_chromosome_type}
\usage{
get_chromosome_type(chromosome, chrID, minRR_object)
}
\arguments{
\item{chromosome}{xx}

\item{chrID}{xx}

\item{minRR_object}{xx}

\item{RR_object}{xxx}
}
\value{
chr_type xx
}
\description{
this function applyes a test based on min(RR) to determine if a chromosome is telocentric (a chromosomal arm) or metacentric (a whole chromosome)
}
