% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_full_name.R
\name{get_genome_full_name}
\alias{get_genome_full_name}
\title{get_genome_full_name}
\usage{
get_genome_full_name(inputGenomeName)
}
\arguments{
\item{inputGenomeName}{: string}
}
\value{
genomeName : string
}
\description{
get the species full name from the input genome short name
}
