% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_of_chromosomes.R
\name{get_list_of_chromosomes}
\alias{get_list_of_chromosomes}
\title{get_list_of_chromosomes}
\usage{
get_list_of_chromosomes(inputData)
}
\arguments{
\item{inputData}{xx}
}
\value{
chrList a vector of IDs of chromosomes
}
\description{
get the list of chromosomes from the given dataset
}
