% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_RR_value_based_on_polynomial.R
\name{get_min_RR_value_based_on_polynomial}
\alias{get_min_RR_value_based_on_polynomial}
\title{get_min_RR_value_based_on_polynomial}
\usage{
get_min_RR_value_based_on_polynomial(chromosome)
}
\arguments{
\item{chromosome}{xx}
}
\value{
minRR_object
}
\description{
........
}
