% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RR_values_for_chromsome.R
\name{get_RR_values_for_chromsome}
\alias{get_RR_values_for_chromsome}
\title{Get local recombination rate values for a chromsome}
\usage{
get_RR_values_for_chromsome(inputGenomeName, inputData, inputChrID,
  plots_path)
}
\arguments{
\item{inputGenomeName}{xx}

\item{inputData}{...}

\item{inputChrID}{xx}

\item{plots_path}{xx}
}
\value{
none ..
}
\description{
get recombination rate values for a specific chromsome and save as csv
}
