% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{load_input_data}
\alias{load_input_data}
\title{load_input_data}
\usage{
load_input_data(genomeName, separator, physicalMapUnit)
}
\arguments{
\item{genomeName}{string of the name of the genome (organism) forwhich the data is provided. This will be used for output and plot file names. Default value is an empty string "".}

\item{separator}{string for the character used for seperating values in the input data file. Default value is the tab separator "\\t". Options : ","  ";"  " ".}

\item{physicalMapUnit}{the measurement unit used for the physical map in the input data file. Default value is megabase pair "Mb". Options: basepair "bp".}
}
\description{
read input dataset from a csv/txt file into a dataframe
}
\seealso{
read.table
}
