% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Generate the interactive plot for recombinantion rate and chromatin boundaries}
\usage{
plot_all(chromosome, RR_object, genomeName, chrID, R2DataFrame2D,
  heteroChromatinBoundaries, swSize, telomeres_boundaries, plots_path,
  chrType_object)
}
\arguments{
\item{chromosome}{xxx}

\item{RR_object}{xxx}

\item{genomeName}{x}

\item{chrID}{x}

\item{R2DataFrame2D}{x}

\item{heteroChromatinBoundaries}{x}

\item{swSize}{x}

\item{telomeres_boundaries}{...}

\item{plots_path}{xx}

\item{chrType_object}{....}
}
\value{
plot_ly object
}
\description{
generate interactive plot for the specified chromosome using plot_ly
}
\details{
the generated plot is saved as png image and the interactive version will open up in the browser as html page
}
\seealso{
\code{link{plolty}} and \code{link{ggplot2}}
}
