% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_as_png.R
\name{save_plot_as_png}
\alias{save_plot_as_png}
\title{Save the generated interactive plot as png image}
\usage{
save_plot_as_png(RR_plot, plots_path, genomeName, chrID)
}
\arguments{
\item{RR_plot}{xx}

\item{plots_path}{xx}

\item{genomeName}{xx}

\item{chrID}{xx}
}
\value{
none
}
\description{
Extract the interactive plot as static image saved in a png file
}
