% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot_figures.R
\name{test_plot_figures}
\alias{test_plot_figures}
\title{testing plot_all for paper figures}
\usage{
test_plot_figures(chromosome, RR_object, genomeName, chrID, R2DataFrame2D,
  heteroChromatinBoundaries, swSize, plots_path)
}
\arguments{
\item{chromosome}{xxx}

\item{RR_object}{xxx}

\item{genomeName}{x}

\item{chrID}{x}

\item{R2DataFrame2D}{x}

\item{heteroChromatinBoundaries}{x}

\item{swSize}{x}

\item{plots_path}{xx}
}
\value{
plot_ly object
}
\description{
generate interactive plot for the specified chromosome using plot_ly
}
\details{
the generated plot is saved as png image and the interactive version will open up in the browser as html page
}
