% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_2arms_to_chromosome.R
\name{transform_2arms_to_chromosome}
\alias{transform_2arms_to_chromosome}
\title{transform_2arms_to_chromosome}
\usage{
transform_2arms_to_chromosome(LeftArm, RightArm, lenghtOfFirstArm)
}
\arguments{
\item{LeftArm}{xx}

\item{RightArm}{xx}

\item{lenghtOfFirstArm}{..}
}
\value{
chromosome
}
\description{
transform two related chromosomal arms to the corresponding chromosome
}
