% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_inputData_to_chromosomes.R
\name{transform_inputData_to_chromosomes}
\alias{transform_inputData_to_chromosomes}
\title{transform_inputData_to_chromosomes}
\usage{
transform_inputData_to_chromosomes(inputData, chrList)
}
\arguments{
\item{inputData}{xxx}

\item{chrList}{xxx}
}
\value{
newInputData
}
\description{
transform inputData to chromosomes including seperated chromosomal arms
}
